requirements.txt
fastapi
uvicorn[standard]
google-generativeai
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
requests
sqlalchemy
python-multipart
pydantic
httpx
fastapi
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-generativeai
pydantic
python-multipart
requests
sift-stack-py
sqlalchemy
uvicorn[standard]
psycopg2-binary
fastapi
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-generativeai
httpx
psycopg2-binary
pydantic
python-multipart
requests
sqlalchemy
uvicorn[standard]
